requirements.txt
fastapi[standard]
uvicorn
nats.py
